+++ Evergreen/Public/Export-EvergreenManifest.ps1
@@ -1,27 +1,6 @@
 Function Export-EvergreenManifest {
     <#
-        .SYNOPSIS
-            Exports an Evergreen application JSON manifest as a hashtable.
-
-        .DESCRIPTION
-            Exports an Evergreen application JSON manifest as a hashtable.
-
-        .NOTES
-            Site: https://stealthpuppy.com
-            Author: Aaron Parker
-            Twitter: @stealthpuppy
-        
-        .LINK
-            https://stealthpuppy.com/Evergreen/
-
-        .PARAMETER Name
-            The application name to return details for. The list of supported applications can be found with Find-EvergreenApp.
-
-        .EXAMPLE
-            Export-EvergreenManifest -Name "MicrosoftEdge"
-
-            Description:
-            Exports the application manifest for the application "MicrosoftEdge".
+        .EXTERNALHELP Evergreen-help.xml
     #>
     [OutputType([System.Management.Automation.PSObject])]
     [CmdletBinding(SupportsShouldProcess = $False, HelpURI = "https://stealthpuppy.com/Evergreen/")]