+++ scripts/in-progress/Get-Opera.ps1
@@ -1,16 +1,27 @@
 # Need to find the correct query to send to the host
+$tempFile = New-TemporaryFile
 $Uri = "https://autoupdate.geo.opera.com/"
 $Headers = @{
     "Connection" = "Keep-Alive"
 }
 
-$params = @{
-    Uri                  = $Uri
-    UserAgent            = "Opera autoupdate agent"
-    UseBasicParsing      = $True
-    ErrorAction          = "SilentlyContinue"
-    Headers              = $Headers
-    SkipCertificateCheck = $True
+try {
+    $params = @{
+        Uri                  = $Uri
+        UserAgent            = "Opera autoupdate agent"
+        Headers              = $Headers
+        SkipCertificateCheck = $True
+        Method               = "Head"
+        SslProtocol          = "Tls13"
+        UseBasicParsing      = $True
+        OutFile              = $tempFile
+        PassThru             = $True
+        ErrorAction          = "SilentlyContinue"
+    }
+    # Call Invoke-WebRequest
+    Invoke-WebRequest @params
 }
-# Call Invoke-WebRequest
-Invoke-WebRequest @params
+catch {
+    Throw $_
+}
+Get-Content -Path $tempFile